backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
httpx==0.28.*
python-multipart==0.0.*
stripe==11.*
pyjwt[crypto]==2.*
python-dotenv==1.*
openpyxl==3.*
pydantic==2.*
slowapi==0.1.*
asyncpg==0.30.*
